requirement.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
numpy==1.17.0
openpyxl==2.6.2
pandas==0.25.0
pkg-resources==0.0.0
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
XlsxWriter==1.1.8
xlwt==1.3.0
